Add files via upload
Uploaded a new version of `CleanGWASData.R` to comply with the `gwas-ssf` validation needed for submission to the GWAS Catalog. Now removes all SNPs not containing an alphabetical character as allele and reformats position to be an integer instead of a numeric (issues will otherwise arise when reading the data into Python via pandas, and attempting to convert to integer, probably due to rounding (really annoying)).